move `find_and_fix_unintialized_timelines` to `timelines` module
