+++ pageserver/src/tenant_mgr.rs
@@ -4,14 +4,13 @@
 use crate::config::PageServerConf;
 use crate::layered_repository::LayeredRepository;
 use crate::pgdatadir_mapping::DatadirTimeline;
-use crate::repository::{Repository, Timeline, TimelineSyncStatusUpdate};
+use crate::repository::{Repository, TimelineSyncStatusUpdate};
 use crate::storage_sync::index::RemoteIndex;
 use crate::storage_sync::{self, LocalTimelineInitStatus, SyncStartupData};
 use crate::tenant_config::TenantConfOpt;
 use crate::thread_mgr;
 use crate::thread_mgr::ThreadKind;
-use crate::timelines::CreateRepo;
-use crate::timelines::{self, init_timeline};
+use crate::timelines::{self, find_and_fix_unintialized_timelines, CreateRepo};
 use crate::walredo::PostgresRedoManager;
 use crate::{DatadirTimelineImpl, RepositoryImpl};
 use anyhow::{bail, Context};
@@ -22,7 +21,6 @@ use std::collections::HashMap;
 use std::fmt;
 use std::sync::Arc;
 use tracing::*;
-use utils::lsn::Lsn;
 
 use utils::zid::{ZTenantId, ZTimelineId};
 
@@ -431,41 +429,6 @@ fn init_local_repository(
     //   and processing of newly downloaded timelines.
     apply_timeline_remote_sync_status_updates(&repo, status_updates)
         .with_context(|| format!("Failed to bootstrap timelines for tenant {tenant_id}"))?;
-    Ok(())
-}
-
-/// Find for broken timelines and try to fix them.
-fn find_and_fix_unintialized_timelines(
-    repo: &LayeredRepository,
-    registration_queue: &[ZTimelineId],
-) -> anyhow::Result<()> {
-    for timeline_id in registration_queue {
-        let inmem_timeline = repo.get_timeline_load(*timeline_id).with_context(|| {
-            format!("Inmem timeline {timeline_id} not found in tenant's repository")
-        })?;
-
-        if inmem_timeline.get_disk_consistent_lsn() == Lsn(0) {
-            info!(
-                "Attempting to fix an uninitialized timeline (id={})",
-                timeline_id
-            );
-
-            // A zero `disk_consistent_lsn` can probably happen because of the failure
-            // when creating a new timeline.
-            // Attempt to fix this "uninitialized" timeline by re-initializing the timeline.
-            let tenant_id = repo.tenant_id();
-            let initdb_path = repo
-                .conf
-                .tenant_path(&tenant_id)
-                .join(format!("tmp-timeline-{}", timeline_id));
-
-            init_timeline(inmem_timeline, initdb_path, *timeline_id, repo).context(format!(
-                "Failed to re-initialize timeline {} for tenant {}",
-                timeline_id, tenant_id
-            ))?;
-        }
-    }
-
     Ok(())
 }
